requirements.txt
easyocr
langdetect
numpy
opencv-python
pillow
pytesseract
streamlit
